Include more checks in the deploy workflow
Looks like workflow_on is not supported by one of the deno actions. So I
can't use it for the deploy workflow to depend on the test workflow.
This introduces some duplication but I care more about not deploying
when tests are failing and that much duplication is fine for me for the
time being.